+++ crates/honk-platform-linux/src/sway.rs
@@ -13,6 +13,10 @@ pub const MAX_AGE: std::time::Duration = std::time::Duration::from_millis(250);
 mod transport;
 #[cfg(target_os = "linux")]
 pub use transport::{Connection, Peer};
+#[cfg(target_os = "linux")]
+mod observer;
+#[cfg(target_os = "linux")]
+pub use observer::Observer;
 
 #[derive(Debug, Clone, Deserialize, PartialEq, Eq)]
 pub struct Version {